+++ NEWS.md
@@ -5,7 +5,9 @@
 
 ### New features
 * New module `tm_t_abnormality_by_worst_grade` was added for the analysis of laboratory test results with highest grade post-baseline. 
+* Enhanced `tm_t_ancova` to include selected parameter(s), visit(s) and the analysis variable in title of the table. 
 * New module `tm_g_lineplot` was added for creating line plots.
+
 
 ## Enhancements
 * Added support for logging with the `logger` package and added info level logs upon initialization of a module.